legalrag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.10.3,<3.0.0
chromadb==0.4.22
sentence-transformers==2.7.0
PyMuPDF==1.23.8
python-dotenv==1.0.0
numpy<2.0
groq>=0.15.0
clerk-backend-api==1.7.0
PyJWT[crypto]>=2.9.0,<3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-docx==1.1.0
